Client app is now successfully loading a file and rendering to JSON
I had a ton of headaches with Jackson, making CheerpJ fail at runtime.
I have sent a test project to them on Gitter; hopefully they find and
fix a defect.  Meanwhile, I have eliminated all the use of ObjectWriter
.writeValueAsString and ObjectMapper.valueToTree().  They seem to have
trouble with doubles.

In doing this, I have probably broken the UnityMesh view of getShapeNode.

I'm still just sending the JSON to stdout.  The next step is to create a
"native" Javascript method to call back and trigger the render actions.

+++ client/cheerpj.bash
@@ -14,39 +14,43 @@ banner() {
 
 banner 'building the distribution'
 pushd ..
-./gradlew distZip
+./gradlew distZip || exit $?
 popd
 
-rm public/*.jar
-rm public/*.jar.js
+if [ "$1" == "" ]; then
+  rm public/*.jar
+  rm public/*.jar.js
+fi
 
 banner 'extracting JAR files'
-unzip ../desktop/build/distributions/vZome-Linux-7.0.dev.zip -d .
-mv vZome-Linux-7.0.dev/lib/jackson-*.jar public
-mv vZome-Linux-7.0.dev/lib/javax.json*.jar public
-mv vZome-Linux-7.0.dev/lib/vecmath*.jar public
-mv vZome-Linux-7.0.dev/lib/core*.jar public
-mv vZome-Linux-7.0.dev/lib/desktop*.jar public
-rm -rf vZome-Linux-7.0.dev/
+unzip ../desktop/build/distributions/vZome-Linux-7.0.dev.zip -d . || exit $?
+mv vZome-Linux-7.0.dev/lib/jackson-*.jar public || exit $?
+mv vZome-Linux-7.0.dev/lib/javax.json*.jar public || exit $?
+mv vZome-Linux-7.0.dev/lib/vecmath*.jar public || exit $?
+mv vZome-Linux-7.0.dev/lib/core*.jar public || exit $?
+mv vZome-Linux-7.0.dev/lib/desktop*.jar public || exit $?
+rm -rf vZome-Linux-7.0.dev/ || exit $?
 
-pushd public
+pushd public || exit $?
 
-banner 'compiling vecmath'
-/Applications/cheerpj/cheerpjfy.py vecmath-1.6.0-final.jar 
-banner 'compiling jackson-annotations'
-/Applications/cheerpj/cheerpjfy.py jackson-annotations-2.9.3.jar 
-banner 'compiling jackson-core'
-/Applications/cheerpj/cheerpjfy.py jackson-core-2.9.5.jar 
-banner 'compiling javax.json'
-/Applications/cheerpj/cheerpjfy.py javax.json-1.0.4.jar 
-banner 'compiling jackson-databind'
-/Applications/cheerpj/cheerpjfy.py --deps jackson-core-2.9.5.jar:jackson-annotations-2.9.3.jar jackson-databind-2.9.5.jar 
+if [ "$1" == "" ]; then
+  banner 'compiling vecmath'
+  /Applications/cheerpj/cheerpjfy.py vecmath-1.6.0-final.jar  || exit $?
+  banner 'compiling jackson-annotations'
+  /Applications/cheerpj/cheerpjfy.py jackson-annotations-2.9.3.jar  || exit $?
+  banner 'compiling jackson-core'
+  /Applications/cheerpj/cheerpjfy.py jackson-core-2.9.5.jar  || exit $?
+  banner 'compiling javax.json'
+  /Applications/cheerpj/cheerpjfy.py javax.json-1.0.4.jar  || exit $?
+  banner 'compiling jackson-databind'
+  /Applications/cheerpj/cheerpjfy.py --deps jackson-core-2.9.5.jar:jackson-annotations-2.9.3.jar jackson-databind-2.9.5.jar  || exit $?
+fi
 
 banner 'compiling core'
-/Applications/cheerpj/cheerpjfy.py --deps jackson-core-2.9.5.jar:jackson-annotations-2.9.3.jar:jackson-databind-2.9.5.jar:javax.json-1.0.4.jar:vecmath-1.6.0-final.jar core-7.0.jar 
+/Applications/cheerpj/cheerpjfy.py --deps jackson-core-2.9.5.jar:jackson-annotations-2.9.3.jar:jackson-databind-2.9.5.jar:javax.json-1.0.4.jar:vecmath-1.6.0-final.jar core-7.0.jar  || exit $?
 
 banner 'compiling desktop'
-/Applications/cheerpj/cheerpjfy.py --deps jackson-core-2.9.5.jar:jackson-annotations-2.9.3.jar:jackson-databind-2.9.5.jar:javax.json-1.0.4.jar:vecmath-1.6.0-final.jar:core-7.0.jar  desktop-7.0.jar 
+/Applications/cheerpj/cheerpjfy.py --deps jackson-core-2.9.5.jar:jackson-annotations-2.9.3.jar:jackson-databind-2.9.5.jar:javax.json-1.0.4.jar:vecmath-1.6.0-final.jar:core-7.0.jar  desktop-7.0.jar  || exit $?
 
 banner 'FINISHED!'
 banner 'Now run "python3 -m http.server 8080"'
